animation.cpp done now im gonna move to animation manager
