+++ source/main.cpp
@@ -1,4 +1,5 @@
-#include <cmath>
+//???
+//#include <cmath>
 #include "BigNgine.h"
 
 void Start()
@@ -12,9 +13,11 @@ void Start()
 			Player->SetDepth(0.0f);
 
 			
-			auto *pAnimation = new BigNgine::AnimationBehaviour();
+			auto *pAnimation = new BigNgine::AnimationBehaviour(.5f);
 			auto *pRendererBehaviour = new BigNgine::TextureRendererBehaviour();
-			pRendererBehaviour->AddTexture("assets/img/1.png");
+//			pRendererBehaviour->AddTexture("assets/img/13.png");
+//          FIXME(imustend): crashes here, if you first add animation and then renderer without adding texture to renderer
+//			TODO(imustend): add animation manager so that we can add multiple animations to the same entity (but how?)
 			pAnimation->AddTexture("assets/img/1.png");
 			pAnimation->AddTexture("assets/img/2.png");
 			pAnimation->AddTexture("assets/img/3.png");
@@ -29,20 +32,21 @@ void Start()
 
 			auto *pMovement = new BigNgine::PlatformerMovementBehaviour();
 
-			Player->AddBehaviour(pAnimation);
 			Player->AddBehaviour(pRendererBehaviour);
+			Player->AddBehaviour(pAnimation);
 			Player->AddBehaviour(pPhysicsBehaviour);
 			Player->AddBehaviour(pMovement);
 
 			// Make Camera follow the Player
 
-			auto *FollowPlayer = new BigNgine::FollowBehaviour(Player, BigNgine::Vector2(50., 50.));
+//			IMPORTANT(imustend): if i remember correctly, adding renderer to camera doesn't work, but i don't remember why
+/*			auto *FollowPlayer = new BigNgine::FollowBehaviour(Player, BigNgine::Vector2(50., 50.));
 			FollowPlayer->lockRotation = true;
 
 			Scene->Camera->AddBehaviour(FollowPlayer);
 
 			// Add a Debug Grid on top of the Camera
-			//										  does not work
+			//does not work
 
 			auto *GridRenderer = new BigNgine::ShaderRendererBehaviour();
 
@@ -53,9 +57,9 @@ void Start()
 
 			Scene->Camera->SetDepth(.5f);
 			//										  FIXME(imustend): renderer on camera doesnt work
-			Scene->Camera->AddBehaviour(GridRenderer);
+			Scene->Camera->AddBehaviour(GridRenderer);*/
 
-			//										  add floor
+			//add floor
 			auto *Floor = new BigNgine::Entity(
 				BigNgine::Vector2(-500.0f, 300.0f), 0.0f,
 				BigNgine::Vector2(800.0f, 40.0f));
@@ -120,8 +124,8 @@ void Start()
 				Scene->Camera,
 				BigNgine::Vector2(-610., -410.));
 
-			Sky->AddBehaviour(SkyRenderer);
 			Sky->AddBehaviour(FollowCamera);
+			Sky->AddBehaviour(SkyRenderer);
 
 			Scene->AddEntity(Floor);
 			Scene->AddEntity(Floor2);
