+++ source/types/texture/texture.cpp
